refactor: Move AccountCreatedEvent to event package and update imports
